text  extraction fixes
