+++ frontend/src/App.css
@@ -1,9 +1,29 @@
 html {
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans', Arial, sans-serif;
+}
+
+body {
+    background-color: rgba(255, 255, 255, 0.064);
 }
 
 .article {
     overflow-y: scroll;
-    min-width: 100%;
-    max-height: 70vh;
+    width: fit-content;
+    max-width: 650px;
+    max-height: 50vh;
+    border-radius: 10px;
+    padding: 5px;
+    margin: 5px;
+    background-color: rgb(255, 255, 255);
+}
+
+.main {
+    width: fit-content;
+    position: sticky;
+    top: 0;
+    right: 0;
+    padding: 10px;
+    margin: 10px;
+    border-radius: 5px;
+    background-color: rgb(226, 226, 226);
 }
