+++ frontend/src/App.tsx
@@ -5,12 +5,11 @@ function App() {
   const [text, setText] = useState([]);
 
   const callExtension = () => {
-    console.log(window.document.body.innerHTML);
     let text = '';
     Array.from(document.body.getElementsByTagName('p')).forEach(
       (element) => (text += element.innerText)
     );
-    console.log(text);
+
     chrome.tabs.query({ active: true, currentWindow: true }, function (tabs) {
       chrome.tabs.sendMessage(
         tabs[0].id,
@@ -22,15 +21,26 @@ function App() {
     });
   };
 
+  const close = () => {
+    setText([]);
+  };
+
   return (
-    <div>
+    <div className="main">
       <h1>TLDR</h1>
-      <button onClick={() => callExtension()}>Extract Article</button>
-      <div className="article">
-        {text.map((t) => (
-          <p>{t}</p>
-        ))}
-      </div>
+
+      {text.length > 0 && (
+        <div className="article">
+          {text.map((t) => (
+            <p>{t}</p>
+          ))}
+        </div>
+      )}
+
+      {!text.length && (
+        <button onClick={() => callExtension()}>Summarize</button>
+      )}
+      {text.length > 0 && <button onClick={() => close()}>close</button>}
     </div>
   );
 }
