+++ frontend/src/index.tsx
@@ -1,9 +1,11 @@
 import ReactDOM from 'react-dom';
-import { MemoryRouter } from 'react-router-dom'
-import AllRoutes from "./AllRoutes"
+import { MemoryRouter } from 'react-router-dom';
+import AllRoutes from './AllRoutes';
+import './index.css';
 
 ReactDOM.render(
-    <MemoryRouter>
-        <AllRoutes />
-    </MemoryRouter>
-    , document.getElementById('root'));
+  <MemoryRouter>
+    <AllRoutes />
+  </MemoryRouter>,
+  document.getElementById('root')
+);
